+++ lib/kitchen/directions/bake_chapter_introductions/v2.rb
@@ -65,6 +65,7 @@ module Kitchen::Directions::BakeChapterIntroductions
         "> :not([data-type='metadata']):not(.splash):not(.has-splash)"
       ).cut
 
+      title.id = "#{introduction_page.id}_titlecreatedbycookbook"
       introduction_page.append(child:
         <<~HTML
           <div class="intro-body">