+++ spec/kitchen_spec/directions/bake_screenreader_spans_spec.rb
@@ -35,7 +35,7 @@ RSpec.describe Kitchen::Directions::BakeScreenreaderSpans do
 
     expect(book1.pages.first).to match_normalized_html(
       <<~HTML
-        <div data-type="page">
+        <div data-type="page" id="testidOne">
           <div>hello <u data-effect="double-underline"><span data-screenreader-only="true">double underline</span>world<span data-screenreader-only="true">end double underline</span></u>. <u data-effect="underline"><span data-screenreader-only="true">underline</span>underlined<span data-screenreader-only="true">end underline</span></u></div>
           <p class="public-domain"><span data-screenreader-only="true">public domain text</span>Proud Immigrant Citizen @primmcit<span data-screenreader-only="true">end public domain text</span></p>
           <p class="student-sample"><span data-screenreader-only="true">student sample text</span>Suddenly,<u data-effect="underline"><span data-screenreader-only="true">underline</span>stood<span data-screenreader-only="true">end underline</span></u> .<span data-screenreader-only="true">end student sample text</span></p>