+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 * Fix bug introduced in #40 by adding ids to introduction pages
+* Remove iframe baking from `BakeAutotitledNotes`
+* Define rex_link on `ElementBase`
+* Link to rex from iframe in `BakeIframes`
 
 ## [v1.5.0] - 2022-4-08